formatId rework and duration calculation overflow fix

+++ src/Album.cpp
@@ -12,7 +12,7 @@ Album::Album(int id,
              QString const& artistName,
              QString const& name,
              int count,
-             int duration,
+             double duration,
              QObject *parent)
     : ItemGroupBase(id, name, count, duration, parent),
       m_artistName(artistName) {
