+++ src/Artist.hpp
@@ -17,7 +17,7 @@ public:
            int downloads,
            int rating,
            int count,
-           int duration,
+           double duration,
            QObject *parent);
     ~Artist();
